refactor(resource): 패키지 이름을 view를 위한 자원임을 더 잘 드러내는 resource로 변경
