+++ src/main/java/subway/controller/LineController.java
@@ -5,8 +5,8 @@ import subway.domain.LineRepository;
 import subway.domain.SectionRepository;
 import subway.view.InputView;
 import subway.view.OutputView;
-import subway.view.text.Function;
-import subway.view.text.Screen;
+import subway.view.resource.Function;
+import subway.view.resource.Screen;
 
 public class LineController {
     public void run(InputView inputView) {
