+++ src/main/java/subway/controller/SectionController.java
@@ -3,8 +3,8 @@ package subway.controller;
 import subway.domain.SectionRepository;
 import subway.view.InputView;
 import subway.view.OutputView;
-import subway.view.text.Function;
-import subway.view.text.Screen;
+import subway.view.resource.Function;
+import subway.view.resource.Screen;
 
 
 public class SectionController {
